Updated class names for table-like structure and image-text alignment
 A slight change was made in the name of the container class for applying columns on text.

To create a table-like structure without using an actual HTML <table> element, use the class name "columnsText" on the <div> element. This class will apply column styling to the text content.

However, if you want to align an image next to text, such as in the example of the header, and justify the content to the center, use the class name "columnsTitle" on the <div> element. This class will handle the image and text alignment as well as centering the content.