Implement Configurations of Events in QDroid
In this epic commit, we unveil the magnificent symphony of event configurations within the realm of QDroid. Embark on a journey guided by Qdler, where events become the lifeblood that fuels specific behaviors, weaving a tapestry of interactions within your application.

This commit introduces the ability to bind class functions to events, introduces the Provide Feedback checkbox, enables the assignment of multiple events to a single class function and bind navigation actions to events